Simple Driving Game
I developed this project as part of the Unity Junior Programming course. The assets and other content belong to the Unity course. In the first scene of the project, we implemented basic vehicle driving and transform functions. The second scene features a flight game coded by me, utilizing assets from the Unity course.